Tests passing on other device

+++ build/index.js
@@ -40,17 +40,19 @@ async function uploadBackend () {
 
     const paths = [
         './server/index.js',
+        './server/index.js.map',
         './server/Dockerfile',
+        './server/package.json',
     ];
 
     await Promise.all(paths.map(async (path) =>
         await $`sshpass -f './sshPass.txt' rsync ${path} ${REMOTE_ADDRESS}:~${REMOTE_BACKEND_PATH}`
     ));
 
-    // upload prod.package.json and rename it to package.json
-    await $`sshpass -f './sshPass.txt' rsync ./server/prod.package.json ${REMOTE_ADDRESS}:~${REMOTE_BACKEND_PATH}/package.json`
-    // upload prod.env and rename it to .env
-    await $`sshpass -f './sshPass.txt' rsync ./server/prod.env ${REMOTE_ADDRESS}:~${REMOTE_BACKEND_PATH}/.env`
+    if (fs.existsSync('./server/prod.env')) {
+        // upload prod.env and rename it to .env
+        await $`sshpass -f './sshPass.txt' rsync ./server/prod.env ${REMOTE_ADDRESS}:~${REMOTE_BACKEND_PATH}/.env`;
+    }
 }
 
 (async () => {
